// app-training-mgmt/fe/src/interfaces/normalizedEmployeeInterface.ts
export interface NormalizedEmployeeInterface {
	id?: string;
	name?: string;
	email?: string;
	doj?: string;
	reportingTo?: string;
	currDesignation?: string;
	currClient1?: string;
	currClient2?: string;
	currClient3?: string;
	currClient4?: string;
	coreTechStack?: string;
	secondaryTechStack?: string;
	status?: string;
}